# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: AI Community Platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /events:
    get:
      operationId: listEvents
      tags: [events]
      summary: List events (cursor-paginated)
      parameters:
        - name: status
          in: query
          schema:
            type: string
            enum: [upcoming, past]
            default: upcoming
        - name: mode
          in: query
          schema:
            type: string
            enum: [online, in_person]
        - name: q
          in: query
          schema:
            type: string
        - name: cursor
          in: query
          schema:
            type: string
        - name: limit
          in: query
          schema:
            type: integer
            default: 12
            maximum: 60
      responses:
        "200":
          description: Cursor-paginated events
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EventListPageResponse"
  /events/upcoming:
    get:
      operationId: getUpcomingEvents
      tags: [events]
      summary: Get upcoming events (dashboard widget)
      parameters:
        - name: limit
          in: query
          schema:
            type: integer
            default: 5
      responses:
        "200":
          description: Upcoming events
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/EventDetail"
  /events/{slug}:
    get:
      operationId: getEvent
      tags: [events]
      summary: Get a single event by slug
      parameters:
        - name: slug
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Event details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EventDetail"
        "404":
          description: Not found
  /events/{slug}/rsvp:
    post:
      operationId: rsvpEvent
      tags: [events]
      summary: RSVP to an event (going / interested / cancelled)
      parameters:
        - name: slug
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RsvpBody"
      responses:
        "200":
          description: RSVP result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EventRsvp"
        "409":
          description: Event full
  /admin/events:
    get:
      operationId: adminListEvents
      tags: [events]
      summary: List all events for admin management
      responses:
        "200":
          description: All events
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/EventDetail"
    post:
      operationId: adminCreateEvent
      tags: [events]
      summary: Create an event (admin)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateEventBody"
      responses:
        "201":
          description: Created event
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EventDetail"
  /admin/events/{slug}:
    patch:
      operationId: adminUpdateEvent
      tags: [events]
      summary: Update an event (admin)
      parameters:
        - name: slug
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateEventBody"
      responses:
        "200":
          description: Updated event
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EventDetail"
    delete:
      operationId: adminDeleteEvent
      tags: [events]
      summary: Delete an event (admin)
      parameters:
        - name: slug
          in: path
          required: true
          schema:
            type: string
      responses:
        "204":
          description: Deleted
  /admin/events/{slug}/cover:
    post:
      operationId: adminUploadEventCover
      tags: [events]
      summary: Upload event cover image (admin)
      parameters:
        - name: slug
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                cover:
                  type: string
                  format: binary
              required: [cover]
      responses:
        "200":
          description: Cover URL
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EventCoverResponse"

  # ── Forum ────────────────────────────────────────────────────────────────────
...
  /storage/objects/{objectPath}:
    get:
      operationId: getStorageObject
      tags: [storage]
      summary: Serve a stored object
      parameters:
        - name: objectPath
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Object file
components:
  schemas:
...
    Event:
      type: object
      properties:
        id:
          type: string
        title:
          type: string
        slug:
          type: string
        description:
          type: string
        startsAt:
          type: string
          format: date-time
        endsAt:
          type: string
          format: date-time
        location:
          type: string
          nullable: true
        capacity:
          type: integer
          nullable: true
        isOnline:
          type: boolean
        meetingUrl:
          type: string
          nullable: true
        coverUrl:
          type: string
          nullable: true
        createdBy:
          type: string
        createdAt:
          type: string
          format: date-time
      required: [id, title, slug, description, startsAt, endsAt, isOnline, createdBy, createdAt]
    EventDetail:
      allOf:
        - $ref: "#/components/schemas/Event"
        - type: object
          properties:
            counts:
              type: object
              properties:
                going:
                  type: integer
                interested:
                  type: integer
              required: [going, interested]
            myRsvp:
              type: string
              enum: [going, interested, cancelled]
              nullable: true
          required: [counts]
    EventListPageResponse:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/EventDetail"
        nextCursor:
          type: string
          nullable: true
      required: [items, nextCursor]
    EventRsvp:
      type: object
      properties:
        id:
          type: string
        eventId:
          type: string
        userId:
          type: string
        status:
          type: string
          enum: [going, interested, cancelled]
        createdAt:
          type: string
          format: date-time
      required: [id, eventId, userId, status, createdAt]
    RsvpBody:
      type: object
      properties:
        status:
          type: string
          enum: [going, interested, cancelled]
      required: [status]
    CreateEventBody:
      type: object
      properties:
        title:
          type: string
        description:
          type: string
        startsAt:
          type: string
          format: date-time
        endsAt:
          type: string
          format: date-time
        location:
          type: string
          nullable: true
        capacity:
          type: integer
          nullable: true
        isOnline:
          type: boolean
        meetingUrl:
          type: string
          nullable: true
        coverUrl:
          type: string
          nullable: true
      required: [title, description, startsAt, endsAt]
    UpdateEventBody:
      type: object
      properties:
        title:
          type: string
        description:
          type: string
        startsAt:
          type: string
          format: date-time
        endsAt:
          type: string
          format: date-time
        location:
          type: string
          nullable: true
        capacity:
          type: integer
          nullable: true
        isOnline:
          type: boolean
        meetingUrl:
          type: string
          nullable: true
        coverUrl:
          type: string
          nullable: true
    EventCoverResponse:
      type: object
      properties:
        coverUrl:
          type: string
      required: [coverUrl]

    # ── Forum ──────────────────────────────────────────────────────────────────
...